Refactored model layers.
