+++ models/fatchord_version.py
@@ -4,7 +4,6 @@ import torch.nn.functional as F
 from utils.distribution import sample_from_discretized_mix_logistic
 from utils.display import *
 from utils.dsp import *
-import os
 import numpy as np
 from pathlib import Path
 from typing import Union
